test: split large test module
